Set calendar page templates on class,
CalendarPage templates were hardcoded into the model methods, which makes overwriting those templates difficult in an app that extends AbstractCalendarPage. While it would be possible to use a template folder like `templates/joyous/*` for templates this may not be reached depending on the order of installed apps. This change moves those template names to properties on the model class for easier customisations and integration with existing apps.